Add constants for phase A from Sanchez-Valle et al., 2008
FIXME: At the moment, it is not clear whether the pressure
derivatives for the ecs, used by the committed shell script,
should be multiplied by two, or used as given in the paper.
More sensible answers are obtained with the latter, but this
will need to be fixed in the future.